==> accidents-prediction/05_preprocess.ipynb <==
# Preprocessing

import time
import pandas as pd
from datetime import datetime
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix
from sklearn.tree import DecisionTreeClassifier
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import MinMaxScaler, StandardScaler
from sklearn.metrics import f1_score, roc_auc_score, confusion_matrix
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from sklearn.svm import SVC
from sklearn.neural_network import MLPClassifier

def evaluate_model(predictions, true_labels):
    accuracy = accuracy_score(true_labels, predictions)
    precision = precision_score(true_labels, predictions, average="weighted")
    recall = recall_score(true_labels, predictions, average="weighted")
    f1 = f1_score(true_labels, predictions, average="weighted")
    confusion = confusion_matrix(true_labels, predictions)
    
    return {
        "accuracy": accuracy,
        "precision": precision,
        "recall": recall,
        "f1": f1,
        "confusion_matrix": confusion
    }

start = time.time()

df = pd.read_csv("./data/accidents_engineered_ordinal.csv")

end = time.time()

#Subtract Start Time from The End Time
total_time = end - start
print(f"Loaded dataset in {round(total_time, 2)}s")

### Standardization (StandardScaler())
- `nbv`
- `lartpc`
- `larrout`
- `vma`
- `occutc`?
- `age`?

columns_to_standardize = ["nbv", "age"]

# Instantiate the StandardScaler
scaler = StandardScaler()

# Fit the scaler on the columns you want to standardize
scaler.fit(df[columns_to_standardize])

# Transform the selected columns to their standardized values
df[columns_to_standardize] = scaler.transform(df[columns_to_standardize])

## Scaling (MinMaxScaler())

columns_to_scale = ["nbv", "age"]

# Instantiate the MinMaxScaler
scaler = MinMaxScaler()

# Fit the scaler on the columns you want to scale
scaler.fit(df[columns_to_scale])

# Transform the selected columns to their scaled values
df[columns_to_scale] = scaler.transform(df[columns_to_scale])

## Drop rows with missing values

df = df.dropna().reset_index(drop=True)

## Encode final variable

mapping = "ordinal"
if mapping == "ordinal":
    grav_order = [1, 2, 3, 4]
    df["grav"] = pd.Categorical(df["grav"], categories=grav_order, ordered=True)
if mapping == "binary":
    df["grav"] = pd.Categorical(df["grav"])

## Convert columns to categorical

cat_columns = ["col", "prof", "plan", "situ", "obs", "choc", "manv",
               "lum", "agg", "catr", "circ", "catv", "catu", "sexe"]

for col in cat_columns:
    df[col] = df[col].astype("category")

## Save

start = time.time()

df.to_csv("./data/accidents_preprocessed.csv", index=False)

end = time.time()

total_time = end - start
print(f"Wrote engineered dataset in {round(total_time, 2)}s")
print(f"There are {len(df):,} rows and {len(df.columns)} columns in the final dataset that will be used for modeling.")

df.info()

now = datetime.now()
current_time = now.strftime("%H:%M:%S")
print(f"Last run at {current_time}")
